% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestModel.R
\name{bestModel}
\alias{bestModel}
\title{The best linear model for a multi-variable problem}
\usage{
bestModel(df, dv)
}
\arguments{
\item{data.frame}{and name of dependant variable}
}
\value{
the object of class 'lm' which is the best fit
}
\description{
The best linear model for a multi-variable problem
}
\examples{
lm_best <- bestModel(mtcars,"mpg")
summary(lm_best)

}
